RAG_Llama/requirements.txt
langchain-groq==0.1.3
langchain==0.1.17
llama-parse==0.1.3
qdrant-client==1.9.1
fastembed==0.2.7
flashrank==0.2.4
unstructured==0.13.6
PyPDF2
flask
python-dotenv
Markdown